First Commit
just started by making the basics by taking all of the info to make the team name and the record into a list. need to make a player class to take the player info and do the same. also need to make the wins, losses, and ties into integers and not strings, want to make them be able to add to the wins and losses.